Working on the "big" div.
I created the $default var for the App class, which has all of the defaults for the cards (and I made the card function use that array, rather than it's hard-coded one).
I made the preview and action span 4 instead of 12, and removed the hack to change the width of .span12.preview and made "big" a span8.
"big" is now tabbable, and has all of the tabs it needs (and their titles are displayed via the __ function).
Made a "reset" for the preview & the print.

Signed-off-by: James Costian <james@jamescostian.com>